+++ postgres/CHANGES.md
@@ -12,6 +12,7 @@
     // execute raw string sql.
     let row_affected: u64 = "SELECT 1; SELECT 1".execute(&client).await?;
     ```
+- remove `Client::pipeline` and `Pool::pipeline`. `pipeline::Pipeline` type can be execute with `Execute::query` method
 - remove `dev::AsParams` trait export. It's not needed for implementing `Query` trait anymore    
 
 ## Change
@@ -31,6 +32,7 @@
     // statement have no value params and can be used for query.
     let stream = stmt.query(&client)?;
     ```
+- rename `Pipeline::query` to `Pipeline::pipe_query` to avoid naming collision with `Execute::query`   
 - `AsyncLendingIterator` is no longer exported from crate's root path. use `iter::AsyncLendingIterator` instead
 - `query::RowStreamOwned` and `row::RowOwned` are no longer behind `compat` crate feature anymore
 - `statement::Statement::unnamed` must bind to value parameters with `bind` or `bind_dyn` before calling `Execute` methods.
@@ -38,9 +40,9 @@
     let stmt = Statement::unnamed("SELECT * FROM users WHERE id = $1", &[Type::INT4]);
     let row_stream = stmt.bind([9527]).query(&client);
     ```
-- `Query::_send_encode_query` method's return type is changed to `Result<(<S as Encode>::Output<'_>, Response), Error>`. Enabling further simplify of the surface level API at the cost of more internal complexity
-- `Encode` trait implementation detail change.
-- `IntoStream` trait is renamed to `IntoResponse`
+- `Query::_send_encode_query` method's return type is changed to `Result<(<S as Encode>::Output, Response), Error>`. Enabling further simplify of the surface level API at the cost of more internal complexity
+- `Encode` trait implementation detail change
+- `IntoStream` trait is renamed to `IntoResponse` with implementation detail change
 
 ## Add
 - add `Execute` trait for extending query customization